core/scrape_orchestrator.py
"""Scraping orchestration logic for One_Touch_Plus.

This module coordinates multi-threaded or asynchronous scraping tasks, 
dynamic content handling, and interaction with various handlers.
It ensures that all parts of the scraping process work together.
"""

import asyncio
import logging

# Import necessary handlers and modules (these are stubbed and will be implemented elsewhere)
from One_Touch_Plus.core.crawl_manager import CrawlManager
from One_Touch_Plus.handlers import dynamic_content_utils, captcha_handler
from One_Touch_Plus.handlers import pdf_text_extractor, image_handler
from One_Touch_Plus.modules import output_reporter, output_validator, proxy_manager

logger = logging.getLogger(__name__)

async def start_scraping(config: dict, target):
    """Start the scraping process for a given target (single URL or list of URLs).

    Args:
        config (dict): The loaded configuration for the scraper.
        target: A single URL (string) or a list of URLs (for batch mode).
    """
    # Initialize components for scraping
    max_depth = config.get('scraper', {}).get('max_depth', None)
    crawl_mgr = CrawlManager(max_depth=max_depth)
    proxy_mgr = proxy_manager.ProxyManager(config.get('proxies'))  # Manage proxies if provided
    reporter = output_reporter.OutputReporter(config)
    validator = output_validator.OutputValidator(config)

    # Determine if target is a single URL or a collection of URLs
    url_list = [target] if isinstance(target, str) else target if isinstance(target, list) else []
    if not url_list:
        logger.error(f"Invalid target type: {type(target)}. Must be str or list of str.")
        return

    # Enqueue initial URLs
    for url in url_list:
        crawl_mgr.add_url(url, depth=0)

    # Primary scraping loop (simplified for scaffold)
    # TODO: Implement concurrency controls (e.g., ThreadPool or asyncio tasks) for efficiency
    while True:
        result = crawl_mgr.get_next_url()
        if not result:
            break  # No more URLs in queue
        current_url, current_depth = result

        logger.info(f"Scraping URL: {current_url}")
        # TODO: Fetch content from current_url (HTTP GET or Selenium driver.get)
        content = ""  # Placeholder for page HTML or content

        # Handle dynamic content (infinite scroll, load more buttons, etc.)
        # dynamic_content_utils.expand_content(driver_or_content, config)

        # Handle CAPTCHA if prompted on this page
        # captcha_handler.handle_captcha(driver_or_content, config)

        # Extract main data from the content
        data = {}  # Placeholder for parsed data (e.g., text, links, etc.)

        # If the page contains PDF links or images, process them
        # Example: text = pdf_text_extractor.extract_text_from_pdf(pdf_path, config)
        # Example: img_text = image_handler.process_image(image_bytes, config)

        # Validate the extracted data
        if not validator.validate(data):
            logger.warning(f"Data from {current_url} failed validation checks.")
            # Decide on continuation or cleanup if validation fails (based on requirements)

        # Generate output report or save data using the reporter
        reporter.generate_report(data)

        # Find new URLs in the content and add to crawl queue if within depth limit
        # TODO: Implement link extraction from content (HTML parsing) and depth check
        # if current_depth < crawl_mgr.max_depth:
        #     for new_url in extract_links(content):
        #         crawl_mgr.add_url(new_url, depth=current_depth + 1)

    logger.info("Scraping process completed.")
    # TODO: Return results or summary if needed for further processing
