requirements.txt

## requirements.txt
```text
PyYAML
schema
requests
aiohttp
pytesseract
selenium
pdf2image
PyMuPDF
